SHTOOLS.jl

The SHTOOLS.jl package wraps SHTOOLS, the Spherical Harmonic Tools.

SHTOOLS is an archive of Fortran 95 software that can be used to perform spherical harmonic transforms, multitaper spectral analyses, expansions of functions into Slepian bases, and standard operations on global gravitational and magnetic field data.

Legendre functions

4π normalized

SHTOOLS.PlmBar! — Function

PlmBar!(p::AbstractVector{Cdouble},
        lmax::Integer,
        z::Union{AbstractFloat,Integer};
        csphase::Integer=1,
        cnorm::Integer=0,
        exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the 4π-normalized associated Legendre functions.

See also: PlmBar, PlmBar_d1!, PlBar!, PlmON!, PlmSchmidt!, PLegendreA!, PlmIndex

SHTOOLS.PlmBar — Function

p = PlmBar(lmax::Integer,
           z::Union{AbstractFloat,Integer};
           csphase::Integer=1,
           cnorm::Integer=0,
           exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the 4π-normalized associated Legendre functions.

See also: PlmBar!, PlmBar_d1, PlBar, PlmON, PlmSchmidt, PLegendreA, PlmIndex

SHTOOLS.PlmBar_d1! — Function

PlmBar_d1!(p::AbstractVector{Cdouble},
           dp::AbstractVector{Cdouble}, 
           lmax::Integer,
           z::Union{AbstractFloat,Integer};
           csphase::Integer=1,
           cnorm::Integer=0,
           exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the 4π-normalized associated Legendre functions and first derivatives.

See also: PlmBar_d1, PlmBar_d1, PlBar_d1!, PlmON_d1!, PlmSchmidt_d1!, PLegendreA_d1!, PlmIndex

SHTOOLS.PlmBar_d1 — Function

p, dp = PlmBar_d1(lmax::Integer,
                  z::Union{AbstractFloat,Integer};
                  csphase::Integer=1,
                  cnorm::Integer=0,
                  exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the 4π-normalized associated Legendre functions and first derivatives.

See also: PlmBar, PlmBar_d1!, PlBar_d1, PlmON_d1, PlmSchmidt_d1, PLegendreA_d1, PlmIndex

SHTOOLS.PlBar! — Function

PlBar!(p::AbstractVector{Cdouble},
       lmax::Integer,
       z::Union{AbstractFloat,Integer};
       exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the 4π-normalized Legendre polynomials.

See also: PlBar, PlBar_d1!, PlmBar!, PlON!, PlSchmidt!, PLegendre!

SHTOOLS.PlBar — Function

p = PlBar(lmax::Integer,
           z::Union{AbstractFloat,Integer};
           exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the 4π-normalized associated Legendre functions.

See also: PlBar!, PlBar_d1, PlmBar, PlON, PlSchmidt, PLegendre

SHTOOLS.PlBar_d1! — Function

PlBar_d1!(p::AbstractVector{Cdouble},
           dp::AbstractVector{Cdouble}, 
           lmax::Integer,
           z::Union{AbstractFloat,Integer};
           exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the 4π-normalized associated Legendre functions and first derivatives.

See also: PlBar_d1, PlBar_d1, PlmBar_d1!, PlON_d1!, PlSchmidt_d1!, PLegendre_d1!

SHTOOLS.PlBar_d1 — Function

p, dp = PlBar_d1(lmax::Integer,
                  z::Union{AbstractFloat,Integer};
                  exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the 4π-normalized associated Legendre functions and first derivatives.

See also: PlBar, PlBar_d1!, PlmBar_d1, PlON_d1, PlSchmidt_d1, PLegendre_d1

Orthonormalized

SHTOOLS.PlmON! — Function

PlmON!(p::AbstractVector{Cdouble},
       lmax::Integer,
       z::Union{AbstractFloat,Integer};
       csphase::Integer=1,
       cnorm::Integer=0,
       exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the orthonormalized associated Legendre functions.

See also: PlmON, PlmON_d1!, PlON!, PlmBar!, PlmSchmidt!, PLegendreA!, PlmIndex

SHTOOLS.PlmON — Function

p = PlmON(lmax::Integer,
          z::Union{AbstractFloat,Integer};
          csphase::Integer=1,
          cnorm::Integer=0,
          exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the orthonormalized associated Legendre functions.

See also: PlmON!, PlmON_d1, PlON, PlmBar, PlmSchmidt, PLegendreA, PlmIndex

SHTOOLS.PlmON_d1! — Function

PlmON_d1!(p::AbstractVector{Cdouble},
          dp::AbstractVector{Cdouble}, 
          lmax::Integer,
          z::Union{AbstractFloat,Integer};
          csphase::Integer=1,
          cnorm::Integer=0,
          exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the orthonormalized associated Legendre functions and first derivatives.

See also: PlmON_d1, PlmON_d1, PlON_d1!, PlmBar_d1!, PlmSchmidt_d1!, PLegendreA_d1!, PlmIndex

SHTOOLS.PlmON_d1 — Function

p, dp = PlmON_d1(lmax::Integer,
                 z::Union{AbstractFloat,Integer};
                 csphase::Integer=1,
                 cnorm::Integer=0,
                 exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the orthonormalized associated Legendre functions and first derivatives.

See also: PlmON, PlmON_d1!, PlON_d1, PlmBar_d1, PlmSchmidt_d1, PLegendreA_d1, PlmIndex

SHTOOLS.PlON! — Function

PlON!(p::AbstractVector{Cdouble},
       lmax::Integer,
       z::Union{AbstractFloat,Integer};
       exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the orthonormalized associated Legendre functions.

See also: PlON, PlON_d1!, PlmON!, PlBar!, PlSchmidt!, PLegendre!

SHTOOLS.PlON — Function

p = PlON(lmax::Integer,
          z::Union{AbstractFloat,Integer};
          exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the orthonormalized associated Legendre functions.

See also: PlON!, PlON_d1, PlmON, PlBar, PlSchmidt, PLegendre

SHTOOLS.PlON_d1! — Function

PlON_d1!(p::AbstractVector{Cdouble},
          dp::AbstractVector{Cdouble}, 
          lmax::Integer,
          z::Union{AbstractFloat,Integer};
          exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the orthonormalized associated Legendre functions and first derivatives.

See also: PlON_d1, PlON_d1, PlmON_d1!, PlBar_d1!, PlSchmidt_d1!, PLegendre_d1!

SHTOOLS.PlON_d1 — Function

p, dp = PlON_d1(lmax::Integer,
                 z::Union{AbstractFloat,Integer};
                 exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the orthonormalized associated Legendre functions and first derivatives.

See also: PlON, PlON_d1!, PlmON_d1, PlBar_d1, PlSchmidt_d1, PLegendre_d1

Schmidt semi-normalized

SHTOOLS.PlmSchmidt! — Function

PlmSchmidt!(p::AbstractVector{Cdouble},
            lmax::Integer,
            z::Union{AbstractFloat,Integer};
            csphase::Integer=1,
            cnorm::Integer=0,
            exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the Schmidt-normalized associated Legendre functions.

See also: PlmSchmidt, PlmSchmidt_d1!, PlSchmidt!, PlmBar!, PlmON!, PLegendreA!, PlmIndex

SHTOOLS.PlmSchmidt — Function

p = PlmSchmidt(lmax::Integer,
               z::Union{AbstractFloat,Integer};
               csphase::Integer=1,
               cnorm::Integer=0,
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the Schmidt-normalized associated Legendre functions.

See also: PlmSchmidt!, PlmSchmidt_d1, PlSchmidt, PlmBar, PlmON, PLegendreA, PlmIndex

SHTOOLS.PlmSchmidt_d1! — Function

PlmSchmidt_d1!(p::AbstractVector{Cdouble},
               dp::AbstractVector{Cdouble}, 
               lmax::Integer,
               z::Union{AbstractFloat,Integer};
               csphase::Integer=1,
               cnorm::Integer=0,
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the Schmidt-normalized associated Legendre functions and first derivatives.

See also: PlmSchmidt_d1, PlmSchmidt_d1, PlSchmidt_d1!, PlmBar_d1!, PlmON_d1!, PLegendreA_d1!, PlmIndex

SHTOOLS.PlmSchmidt_d1 — Function

p, dp = PlmSchmidt_d1(lmax::Integer,
                      z::Union{AbstractFloat,Integer};
                      csphase::Integer=1,
                      cnorm::Integer=0,
                      exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the Schmidt-normalized associated Legendre functions and first derivatives.

See also: PlmSchmidt, PlmSchmidt_d1!, PlSchmidt_d1, PlmBar_d1, PlmON_d1, PLegendreA_d1, PlmIndex

SHTOOLS.PlSchmidt! — Function

PlSchmidt!(p::AbstractVector{Cdouble},
            lmax::Integer,
            z::Union{AbstractFloat,Integer};
            exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the Schmidt-normalized associated Legendre functions.

See also: PlSchmidt, PlSchmidt_d1!, PlmSchmidt!, PlBar!, PlON!, PLegendre!

SHTOOLS.PlSchmidt — Function

p = PlSchmidt(lmax::Integer,
               z::Union{AbstractFloat,Integer};
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the Schmidt-normalized associated Legendre functions.

See also: PlSchmidt!, PlSchmidt_d1, PlmSchmidt, PlBar, PlON, PLegendre

SHTOOLS.PlSchmidt_d1! — Function

PlSchmidt_d1!(p::AbstractVector{Cdouble},
               dp::AbstractVector{Cdouble}, 
               lmax::Integer,
               z::Union{AbstractFloat,Integer};
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the Schmidt-normalized associated Legendre functions and first derivatives.

See also: PlSchmidt_d1, PlSchmidt_d1, PlmSchmidt_d1!, PlBar_d1!, PlON_d1!, PLegendre_d1!

SHTOOLS.PlSchmidt_d1 — Function

p, dp = PlSchmidt_d1(lmax::Integer,
                      z::Union{AbstractFloat,Integer};
                      exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the Schmidt-normalized associated Legendre functions and first derivatives.

See also: PlSchmidt, PlSchmidt_d1!, PlmSchmidt_d1, PlBar_d1, PlON_d1, PLegendre_d1

Unnormalized

SHTOOLS.PLegendreA! — Function

PLegendreA!(p::AbstractVector{Cdouble},
            lmax::Integer,
            z::Union{AbstractFloat,Integer};
            csphase::Integer=1,
            exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the unnormalized associated Legendre functions.

See also: PLegendreA, PLegendreA_d1!, PLegendre!, PlmBar!, PlmON!, PlmSchmidt!, PlmIndex

SHTOOLS.PLegendreA — Function

p = PLegendreA(lmax::Integer,
               z::Union{AbstractFloat,Integer};
               csphase::Integer=1,
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the unnormalized associated Legendre functions.

See also: PLegendreA!, PLegendreA_d1, PLegendre, PlmBar, PlmON, PlmSchmidt, PlmIndex

SHTOOLS.PLegendreA_d1! — Function

PLegendreA_d1!(p::AbstractVector{Cdouble},
               dp::AbstractVector{Cdouble}, 
               lmax::Integer,
               z::Union{AbstractFloat,Integer};
               csphase::Integer=1,
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the unnormalized associated Legendre functions and first derivatives.

See also: PLegendreA_d1, PLegendreA_d1, PLegendre_d1!, PlmBar_d1!, PlmON_d1!, PlmSchmidt_d1!, PlmIndex

SHTOOLS.PLegendreA_d1 — Function

p, dp = PLegendreA_d1(lmax::Integer,
                        z::Union{AbstractFloat,Integer};
                        csphase::Integer=1,
                        exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the unnormalized associated Legendre functions and first derivatives.

See also: PLegendreA, PLegendreA_d1!, PLegendre_d1, PlmBar_d1, PlmON_d1, PlmSchmidt_d1, PlmIndex

SHTOOLS.PLegendre! — Function

PLegendre!(p::AbstractVector{Cdouble},
            lmax::Integer,
            z::Union{AbstractFloat,Integer};
            exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the unnormalized associated Legendre functions.

See also: PLegendre, PLegendre_d1!, PLegendreA!, PlBar!, PlON!, PlSchmidt!

SHTOOLS.PLegendre — Function

p = PLegendre(lmax::Integer,
               z::Union{AbstractFloat,Integer};
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}

Compute all the unnormalized associated Legendre functions.

See also: PLegendre!, PLegendre_d1, PLegendreA, PlBar, PlON, PlSchmidt

SHTOOLS.PLegendre_d1! — Function

PLegendre_d1!(p::AbstractVector{Cdouble},
               dp::AbstractVector{Cdouble}, 
               lmax::Integer,
               z::Union{AbstractFloat,Integer};
               exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Compute all the unnormalized associated Legendre functions and first derivatives.

See also: PLegendre_d1, PLegendre_d1, PLegendreA_d1!, PlBar_d1!, PlON_d1!, PlSchmidt_d1!

SHTOOLS.PLegendre_d1 — Function

p, dp = PLegendre_d1(lmax::Integer,
                        z::Union{AbstractFloat,Integer};
                        exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
p::Vector{Cdouble}
dp::Vector{Cdouble}

Compute all the unnormalized associated Legendre functions and first derivatives.

See also: PLegendre, PLegendre_d1!, PLegendre_d1!, PlBar_d1, PlON_d1, PlSchmidt_d1

Utilities

SHTOOLS.PlmIndex — Function

index = PlmIndex(l::Integer,
                 m::Integer)
index::Int

Compute the index of an array of Legendre functions corresponding to degree l and angular order m.

Spherical harmonic transforms

Equally sampled (N×N) and equally spaced (N×2N) grids

Missing docstring.

Missing docstring for SHExpandDH. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGridDH. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SHExpandDHC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGridDHC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGradientDH. Check Documenter's build log for details.

Gauss-Legendre quadrature grids

Missing docstring.

Missing docstring for SHGLQ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SHExpandGLQ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGridGLQ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SHExpandGLQC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGridGLQC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GLQGridCoord. Check Documenter's build log for details.

Other routines

Missing docstring.

Missing docstring for SHExpandLSQ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGrid2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGridPoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MakeGridPointC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SHMultiply. Check Documenter's build log for details.

Spherical harmonic I/O, storage, and conversions

Spherical harmonic storage

SHTOOLS.SHCilmToVector! — Function

SHCilmToVector!(vector::AbstractVector{Cdouble},
                cilm::AbstractArray{Cdouble,3},
                lmax::Integer;
                exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Convert a three-dimensional array of real spherical harmonic coefficients to a 1-dimensional indexed vector.

See also: SHCilmToVector

SHTOOLS.SHCilmToVector — Function

vector = SHCilmToVector(cilm::AbstractArray{Cdouble,3},
                        lmax::Integer;
                        exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
vector::Vector{Cdouble}

Convert a three-dimensional array of real spherical harmonic coefficients to a 1-dimensional indexed array.

See also: SHCilmToVector!

SHTOOLS.SHVectorToCilm! — Function

SHVectorToCilm!(cilm::AbstractArray{Cdouble,3},
                vector::AbstractVector{Cdouble},
                lmax::Integer;
                exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)

Convert a 1-dimensional indexed vector of real spherical harmonic coefficients to a 3-dimensional array.

See also: SHVectorToCilm

SHTOOLS.SHVectorToCilm — Function

cilm = SHVectorToCilm(cilm::AbstractArray{Cdouble,3},
                      vector::AbstractVector{Cdouble},
                      lmax::Integer;
                      exitstatus::Union{Nothing,Ref{<:Integer}}=nothing)
cilm::Array{Cdouble,3}

Convert a 1-dimensional indexed vector of real spherical harmonic coefficients to a 3-dimensional array.

See also: SHVectorToCilm!

Spherical harmonic conversions